+++ Automation/2_2_1_6_kopioi_tiedot_arviointityökirjaan_ryhmien_tiedot-välilehdelle.robot
@@ -2,15 +2,13 @@
 Library    Process
 
 *** Variables ***
-${EXCEL_FILE}    D:\\Users\\Sonjis\\PycharmProjects\\opetusTukiAutomaatio5\\Data\\siirto.xlsx
+${EXCEL_FILE}    ../Data/arviointityokirja.xlsx  # Adjust path as necessary
 
-${PYTHON_SCRIPT}    D:\\Users\\Sonjis\\PycharmProjects\\opetusTukiAutomaatio5\\Resources\\excel.py
 *** Test Cases ***
 Move Data Using Python Script
     [Documentation]    Moves data from the first sheet to a new Sheet2 using a Python script.
-    Log    Running script with file: ${EXCEL_FILE}
-    ${result}=    Run Process    python    ${PYTHON_SCRIPT}    ${EXCEL_FILE}
-    Log    Script output: ${result.stdout}
-    Log    Script errors: ${result.stderr}
+    ${result}=    Run Process    python    ../Resources/2_2_1_6_kopioi_tiedot_arviointityökirjaan_ryhmien_tiedot-välilehdelle.py   ${EXCEL_FILE}
+    Log    ${result.stdout}
+    Log    ${result.stderr}
     Should Not Contain    ${result.stdout}    Traceback    Script execution should not error
-    Should Not Contain    ${result.stderr}    Traceback    There should be no errors in stderr
+    Should Not Contain    ${result.stderr}    Traceback    There should be no errors in stderr
