+++ Automation/2_2_2_4_Avaa_arviointiexceli.robot
@@ -1,13 +1,6 @@
 *** Settings ***
-Library    RPA.Excel.Files
-
-*** Variables ***
-${EXCEL_FILE}    ../Resources/siirto.xlsx
+Library    Process
 
 *** Test Cases ***
-Open Excel File
-    [Documentation]    Tämä avaa Excel-tiedoston nimeltä siirto.xlsx Resources-kansiosta.
-    Open Workbook    ${EXCEL_FILE}
-    ${sheet_names}=    List Worksheets
-    Log    Tiedoston taulut: ${sheet_names}
-    Close Workbook
+Open Excel With Specific File
+    Start Process    C:\\Program Files\\Microsoft Office\\root\\Office16\\EXCEL.EXE    D:\\Users\\sonjis\\PycharmProjects\\opetusTukiAutomaatio5\\Data\\arviointityokirja.xlsx
